+++ packages/analytics-core/test/network-observer.test.ts
@@ -1,4 +1,10 @@
-import { NetworkEventCallback, NetworkRequestEvent, networkObserver } from '../src/index';
+import {
+  FORBIDDEN_HEADERS,
+  NetworkEventCallback,
+  NetworkRequestEvent,
+  SAFE_HEADERS,
+  networkObserver,
+} from '../src/index';
 import { NetworkObserver } from '../src/network-observer';
 import {
   FetchRequestBody,
@@ -104,18 +110,18 @@ describe('NetworkObserver', () => {
         status: 200,
       });
       expect(events[0].duration).toBeGreaterThanOrEqual(0);
-      expect(events[0].requestWrapper?.headers([], true)).toEqual({
+      expect(events[0].requestWrapper?.headers([...SAFE_HEADERS])).toEqual({
         'Content-Type': 'application/json',
       });
       // expect headers to throw an error if consumed
-      expect(() => events[0].requestWrapper?.headers([], true)).toThrow(TypeError);
+      expect(() => events[0].requestWrapper?.headers([...SAFE_HEADERS])).toThrow(TypeError);
       const expectedResponseHeaders = {
         'content-type': 'application/json',
         'content-length': '20',
         server: 'test-server',
       };
-      expect(events[0].responseWrapper?.headers([], true)).toEqual(expectedResponseHeaders);
-      expect(() => events[0].responseWrapper?.headers([], true)).toThrow(TypeError);
+      expect(events[0].responseWrapper?.headers([...SAFE_HEADERS])).toEqual(expectedResponseHeaders);
+      expect(() => events[0].responseWrapper?.headers([...SAFE_HEADERS])).toThrow(TypeError);
     });
 
     it('should track successful fetch requests with headers (uses Headers object)', async () => {
@@ -151,11 +157,11 @@ describe('NetworkObserver', () => {
         status: 200,
       });
       expect(events[0].duration).toBeGreaterThanOrEqual(0);
-      expect(events[0].requestWrapper?.headers(['safe-header'], true)).toEqual({
+      expect(events[0].requestWrapper?.headers(['safe-header', ...SAFE_HEADERS])).toEqual({
         'content-type': 'application/json',
         'safe-header': 'safe-value',
       });
-      expect(events[0].responseWrapper?.headers([], true)).toEqual({
+      expect(events[0].responseWrapper?.headers([...SAFE_HEADERS])).toEqual({
         'content-type': 'application/json',
         'content-length': '20',
         server: 'test-server',
@@ -260,9 +266,9 @@ describe('NetworkObserver', () => {
         startTime: expect.any(Number),
         endTime: expect.any(Number),
         timestamp: expect.any(Number),
-        requestHeaders: networkRequestEvent.requestWrapper?.headers([], true),
+        requestHeaders: networkRequestEvent.requestWrapper?.headers([...SAFE_HEADERS]),
         requestBodySize: networkRequestEvent.requestWrapper?.bodySize,
-        responseHeaders: networkRequestEvent.responseWrapper?.headers([], true),
+        responseHeaders: networkRequestEvent.responseWrapper?.headers([...SAFE_HEADERS]),
         responseBodySize: networkRequestEvent.responseWrapper?.bodySize,
       });
     });
@@ -574,6 +580,15 @@ describe('NetworkObserver', () => {
     });
 
     describe('headers() should return', () => {
+      it('{} when allowlist is empty', () => {
+        const requestWrapper = new RequestWrapperFetch({
+          headers: {
+            'Content-Type': 'application/fake',
+            'Content-Length': '100',
+          },
+        } as RequestInitSafe);
+        expect(requestWrapper.headers([])).toEqual({});
+      });
       it('an object when headers is an array', () => {
         const requestWrapper = new RequestWrapperFetch({
           headers: [
@@ -581,16 +596,16 @@ describe('NetworkObserver', () => {
             ['Content-Length', '1234'],
           ],
         } as RequestInitSafe);
-        expect(requestWrapper.headers([], true)).toEqual({
+        expect(requestWrapper.headers([...SAFE_HEADERS])).toEqual({
           'Content-Type': 'application/fake',
           'Content-Length': '1234',
         });
       });
-      it('undefined when headers is undefined', () => {
+      it('{} when headers is undefined', () => {
         const requestWrapper = new RequestWrapperFetch({
           headers: undefined,
         } as RequestInitSafe);
-        expect(requestWrapper.headers([], true)).toBeUndefined();
+        expect(requestWrapper.headers()).toEqual({});
       });
     });
 
@@ -627,10 +642,16 @@ describe('NetworkObserver', () => {
     });
 
     describe('.json()', () => {
-      test('json should return {} if no rules are set', async () => {
+      test('should return null if allowlist and exclude are undefined', async () => {
+        const responseWrapper = new ResponseWrapperFetch(mockResponse as unknown as Response);
+        const json = await responseWrapper.json();
+        expect(json).toEqual(null);
+      });
+
+      test('json should return null if no rules are set', async () => {
         const responseWrapper = new ResponseWrapperFetch(mockResponse as unknown as Response);
         const json = await responseWrapper.json([], []);
-        expect(json).toEqual({});
+        expect(json).toEqual(null);
       });
 
       test('json should include allowed fields in response', async () => {
@@ -668,6 +689,13 @@ describe('NetworkObserver', () => {
       });
     });
 
+    describe('.headers()', () => {
+      test('should return {} if allowlist is empty', () => {
+        const responseWrapper = new ResponseWrapperFetch(mockResponse as unknown as Response);
+        expect(responseWrapper.headers()).toEqual({});
+      });
+    });
+
     test('text should return null if text returns nothing', async () => {
       const mockResponseNonJson = {
         ...mockResponse,
@@ -676,7 +704,7 @@ describe('NetworkObserver', () => {
       };
       const responseWrapper = new ResponseWrapperFetch(mockResponseNonJson as unknown as Response);
       const json = await responseWrapper.text();
-      await responseWrapper.text();
+      expect(responseWrapper.text()).rejects.toThrow(TypeError);
       expect(json).toBeNull();
     });
 
@@ -789,7 +817,23 @@ describe('RequestWrapperXhr', () => {
     expect(requestWrapper.body).toBeNull();
   });
 
+  describe('.headers()', () => {
+    test('should return undefined if allowlist empty', () => {
+      const requestWrapper = new RequestWrapperXhr('Hello World!', {});
+      expect(requestWrapper.headers()).toEqual({});
+    });
+  });
+
   describe('.json()', () => {
+    test('should return null if allowlist and exclude are empty', async () => {
+      const requestWrapper = new RequestWrapperXhr(
+        JSON.stringify({ message: 'Hello from mock!', secret: 'secret' }),
+        {},
+      );
+      const json = await requestWrapper.json();
+      expect(json).toEqual(null);
+    });
+
     test('should parse body as JSON', async () => {
       const requestWrapper = new RequestWrapperXhr(
         JSON.stringify({ message: 'Hello from mock!', secret: 'secret' }),
@@ -797,6 +841,7 @@ describe('RequestWrapperXhr', () => {
       );
       const json = await requestWrapper.json(['message', 'secret'], ['secret']);
       expect(json).toEqual({ message: 'Hello from mock!' });
+      expect(requestWrapper.json(['message', 'secret'], ['secret'])).rejects.toThrow(TypeError);
     });
   });
 });
@@ -891,12 +936,12 @@ describe('observeXhr', () => {
           expect(event.type).toBe('xhr');
           expect(event.method).toBe('GET');
           expect(event.url).toBe('https://api.example.com/data');
-          expect(event.responseWrapper?.headers([], true)).toEqual({
+          expect(event.responseWrapper?.headers([...SAFE_HEADERS])).toEqual({
             'Content-Type': 'application/json',
             'Content-Length': '1234',
           });
           expect(event.responseWrapper?.bodySize).toBe(1234);
-          expect(event.requestWrapper?.headers([], true)).toEqual({});
+          expect(event.requestWrapper?.headers([])).toEqual({});
           expect(event.requestWrapper?.bodySize).toBe('hello world!'.length);
           expect(event.duration).toBeGreaterThanOrEqual(0);
           expect(event.startTime).toBeGreaterThanOrEqual(0);
@@ -921,15 +966,39 @@ describe('observeXhr', () => {
 });
 
 describe('ResponseWrapperXhr', () => {
-  test('should return undefined if headersString is empty', async () => {
-    const responseWrapper = new ResponseWrapperXhr(200, '', 0, 'some text');
-    expect(responseWrapper.headers()).toEqual(undefined);
+  describe('.headers()', () => {
+    test('should return {} if allowlist is empty', () => {
+      const responseWrapper = new ResponseWrapperXhr(200, '', 0, 'some text');
+      expect(responseWrapper.headers()).toEqual({});
+    });
+
+    test('should return {} if headerString is empty', () => {
+      const responseWrapper = new ResponseWrapperXhr(200, '', 0, '');
+      expect(responseWrapper.headers()).toEqual({});
+    });
+  });
+
+  test('should return {} if headersString is empty', async () => {
+    const responseWrapper = new ResponseWrapperXhr(200, 'hello=world', 0, 'some text');
+    expect(responseWrapper.headers()).toEqual({});
     expect(() => responseWrapper.headers()).toThrow(TypeError);
     const text = await responseWrapper.text();
     expect(text).toBe('some text');
   });
 
   describe('.json()', () => {
+    test('should return null if allowlist and exclude are empty', async () => {
+      const responseWrapper = new ResponseWrapperXhr(200, '', 0, 'some text');
+      const json = await responseWrapper.json();
+      expect(json).toEqual(null);
+    });
+
+    test('should return null if allowlist is empty', async () => {
+      const responseWrapper = new ResponseWrapperXhr(200, '', 0, 'some text');
+      const json = await responseWrapper.json([], ['secret']);
+      expect(json).toEqual(null);
+    });
+
     test('should parse JSON response', async () => {
       const response = {
         message: 'Hello from mock!',
@@ -1014,6 +1083,26 @@ describe('RequestWrapperFetch', () => {
   });
 
   describe('.json()', () => {
+    test('should return null if allowlist and exclude are empty', async () => {
+      const requestWrapper = new RequestWrapperFetch({
+        body: JSON.stringify({ message: 'Hello from mock!', secret: 'secret' }),
+        headers: new Headers({ 'Content-Type': 'application/json', 'Content-Length': '1234' }),
+        status: 200,
+      } as unknown as RequestInitSafe);
+      const json = await requestWrapper.json();
+      expect(json).toEqual(null);
+    });
+
+    test('should return null if allowlist is empty', async () => {
+      const requestWrapper = new RequestWrapperFetch({
+        body: JSON.stringify({ message: 'Hello from mock!', secret: 'secret' }),
+        headers: new Headers({ 'Content-Type': 'application/json', 'Content-Length': '1234' }),
+        status: 200,
+      } as unknown as RequestInitSafe);
+      const json = await requestWrapper.json([], ['secret']);
+      expect(json).toEqual(null);
+    });
+
     test('should parse body as JSON', async () => {
       const requestWrapper = new RequestWrapperFetch({
         body: JSON.stringify({ message: 'Hello from mock!', secret: 'secret' }),
@@ -1022,7 +1111,7 @@ describe('RequestWrapperFetch', () => {
       } as unknown as RequestInitSafe);
       const json = await requestWrapper.json(['message'], ['secret']);
       expect(json).toEqual({ message: 'Hello from mock!' });
-      expect(requestWrapper.json(['message'], ['secret'])).rejects.toThrow(TypeError);
+      expect(requestWrapper.json(['message', 'secret'], ['secret'])).rejects.toThrow(TypeError);
     });
   });
 });
@@ -1041,30 +1130,25 @@ describe('NetworkRequestEvent', () => {
 });
 
 describe('pruneHeaders', () => {
+  test('should be empty object if allowlist is undefined', () => {
+    const headers = {
+      'Content-Type': 'application/json',
+      'Content-Length': '1234',
+      authorization: 'secretpassword!',
+    };
+    const prunedHeaders = pruneHeaders(headers, { allow: undefined });
+    expect(prunedHeaders).toEqual({});
+  });
   test('should exclude headers that are forbidden', () => {
     const headers = {
       'Content-Type': 'application/json',
       'Content-Length': '1234',
       authorization: 'secretpassword!',
     };
-    const prunedHeaders = pruneHeaders(headers, { allow: [], captureSafeHeaders: true });
+    const prunedHeaders = pruneHeaders(headers, { allow: [...SAFE_HEADERS] });
     expect(prunedHeaders).toEqual({
       'Content-Type': 'application/json',
       'Content-Length': '1234',
-    });
-  });
-
-  test('should not exclude headers if exclude is not provided', () => {
-    const headers = {
-      'Content-Type': 'application/json',
-      'Content-Length': '1234',
-      'Random-Header': 'random-value',
-    };
-    const prunedHeaders = pruneHeaders(headers, { allow: ['Random-Header'], captureSafeHeaders: true });
-    expect(prunedHeaders).toEqual({
-      'Content-Type': 'application/json',
-      'Content-Length': '1234',
-      'Random-Header': 'random-value',
     });
   });
 
@@ -1075,25 +1159,9 @@ describe('pruneHeaders', () => {
       'X-Custom-Header': 'customvalue',
       authorization: 'secretpassword!',
     };
-    const prunedHeaders = pruneHeaders(headers, { allow: [], captureSafeHeaders: true });
+    const prunedHeaders = pruneHeaders(headers, { allow: [...SAFE_HEADERS] });
     expect(prunedHeaders).toEqual({
       'Content-Type': 'application/json',
-      'Content-Length': '1234',
-    });
-  });
-
-  test('should exclude headers and include headers', () => {
-    const headers = {
-      'Content-Type': 'application/json',
-      'Content-Length': '1234',
-      'X-Custom-Header': 'customvalue',
-      authorization: 'secretpassword!',
-    };
-    const prunedHeaders = pruneHeaders(headers, {
-      exclude: ['Content-Type'],
-      allow: ['Content-Type', 'Content-Length'],
-    });
-    expect(prunedHeaders).toEqual({
       'Content-Length': '1234',
     });
   });
@@ -1109,7 +1177,7 @@ describe('pruneHeaders', () => {
     expect(prunedHeaders).toEqual({});
   });
 
-  test('should delete headers if strategy is remove', () => {
+  test('should delete forbidden headers if strategy is remove', () => {
     const headers = {
       'Content-Type': 'application/json',
       'Content-Length': '1234',
@@ -1117,26 +1185,12 @@ describe('pruneHeaders', () => {
       authorization: 'secretpassword!',
     };
     const prunedHeaders = pruneHeaders(headers, {
-      captureSafeHeaders: true,
-      allow: [],
-      exclude: ['Content-Type'],
+      allow: [...SAFE_HEADERS],
       strategy: PRUNE_STRATEGY.REMOVE,
     });
     expect(prunedHeaders).toEqual({
       'Content-Length': '1234',
-    });
-  });
-
-  test('should default allow to be empty array', () => {
-    const headers = {
       'Content-Type': 'application/json',
-      'Content-Length': '1234',
-      authorization: 'secretpassword!',
-    };
-    const prunedHeaders = pruneHeaders(headers, { captureSafeHeaders: true });
-    expect(prunedHeaders).toEqual({
-      'Content-Type': 'application/json',
-      'Content-Length': '1234',
     });
   });
 
@@ -1145,19 +1199,27 @@ describe('pruneHeaders', () => {
       'Content-Type': 'application/json',
       'Content-Length': '1234',
       authorization: 'secretpassword!',
+      cookie: 'secretcookie!',
+      'set-cookie': 'secretsetcookie!',
       'X-Custom-Header': 'customvalue',
+      'Unallowed-Header': 'unallowed-value',
     };
     const prunedHeaders = pruneHeaders(headers, {
-      captureSafeHeaders: true,
-      allow: [],
-      exclude: ['Content-Type'],
+      allow: [...SAFE_HEADERS, 'X-Custom-Header'],
       strategy: PRUNE_STRATEGY.REDACT,
     });
     expect(prunedHeaders).toEqual({
-      'Content-Type': '[REDACTED]',
+      'Content-Type': 'application/json',
       'Content-Length': '1234',
       authorization: '[REDACTED]',
-      'X-Custom-Header': '[REDACTED]',
+      cookie: '[REDACTED]',
+      'set-cookie': '[REDACTED]',
+      'X-Custom-Header': 'customvalue',
+      'Unallowed-Header': '[REDACTED]',
     });
+
+    for (const forbiddenHeader of FORBIDDEN_HEADERS) {
+      expect(prunedHeaders[forbiddenHeader.toLowerCase()]).toBe('[REDACTED]');
+    }
   });
 });